Moved FileWriter from pipeline.io.arches to pipeline.io.file.
